LayoutTests/fast/scrolling/mac/momentum-animator-maybegin-stops-expected.txt

Momentum event reached main thread
PASS scrollEventCount > 0 is true
PASS scrollEventCount is scrollEventCountAtMayBegin
PASS successfullyParsed is true

TEST COMPLETE
